continue code reorganization
